Lots of fixes, faster function eval

+++ header.h
@@ -57,7 +57,7 @@ void compute_noise_wavelet (float*** pol, float*** noise, int nnu, int nk, int n
 unsigned int powerof2 (unsigned int n);
 
 /* fit.c */
-int fit_peak (struct params* p, float *freq, float *amp, float *width, float*** pol, float*** noise, float delta_nu, float delta_k, int nnu, int ntheta, int k);
+int fit_peak (struct params* p, double *freq, double *amp, double *width, float*** pol, float*** noise, float delta_nu, float delta_k, int nnu, int ntheta, int k);
 int funk_single(int m, int n, double* p, double* deviates, double**derivs, void* private_data);
 int fit_back (struct params* p, double* amp, double* cutoff, double* power, float*** pol, float*** noise, float delta_nu, int nnu, int ntheta, int k);
 int funk_back(int m, int n, double* p, double* deviates, double**derivs, void* private_data);
@@ -66,6 +66,6 @@ int funk_back(int m, int n, double* p, double* deviates, double**derivs, void* p
 void output_debug (struct params* p, float***pol, float***noise, int ntheta, int nk, int nnu, int k, int m, int n, double* x, float delta_nu, float delta_k);
 void output_covar (double* covar, int n, struct params* p);
 int read_fits_file (float**** spec, float**** noise, struct params* p, 
-		int* ntheta, int* nk, int* nnu, float* delta_k, float* delta_nu);
+		int* ntheta, int* nk, int* nnu, double* delta_k, double* delta_nu);
 void read_param_file (char* fname, struct params* p);
 void trim (char* str);
